Canonicalize dashboard IA
