+++ app/dashboard/_components/sites-nav.test.tsx
@@ -81,7 +81,7 @@ describe("SitesNav", () => {
       />,
     );
 
-    for (const label of ["Configuration", "Pages", "Consistency", "Overrides", "Admin"]) {
+    for (const label of ["Workspace", "Pages & crawl", "Translation rules", "Settings"]) {
       const link = screen.getByRole("link", { name: label });
       expect(link.getAttribute("data-prefetch")).toBe("false");
     }
