+++ app/dashboard/_components/sites-nav.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { usePathname } from "next/navigation";
-import { ChevronDown, Globe } from "lucide-react";
+import { ChevronDown } from "lucide-react";
 import { useState } from "react";
 
 import {
@@ -27,18 +27,9 @@ export function SitesNav({ sites }: SitesNavProps) {
   const pathname = usePathname();
   const { state } = useSidebar();
   const isCollapsed = state === "collapsed";
-  const allSitesActive = pathname === "/dashboard/sites";
 
   return (
     <SidebarMenu>
-      <SidebarMenuItem>
-        <SidebarMenuButton asChild isActive={allSitesActive} tooltip="All sites">
-          <Link href="/dashboard/sites" aria-current={allSitesActive ? "page" : undefined}>
-            <Globe className="h-4 w-4 shrink-0" />
-            <span className="truncate group-data-[collapsible=icon]:sr-only">All sites</span>
-          </Link>
-        </SidebarMenuButton>
-      </SidebarMenuItem>
       {isCollapsed ? null : sites.length === 0 ? (
         <SidebarMenuItem className="px-2 py-2 text-xs text-sidebar-foreground/70">
           No sites yet.
@@ -65,11 +56,10 @@ function SiteNavItem({ site, pathname }: SiteNavItemProps) {
   const menuId = `site-nav-${site.id}`;
 
   const subItems = [
-    { href: baseHref, label: "Configuration" },
-    { href: `${baseHref}/pages`, label: "Pages" },
-    { href: `${baseHref}/consistency`, label: "Consistency" },
-    { href: `${baseHref}/overrides`, label: "Overrides" },
-    { href: `${baseHref}/admin`, label: "Admin" },
+    { href: baseHref, label: "Workspace" },
+    { href: `${baseHref}/pages`, label: "Pages & crawl" },
+    { href: `${baseHref}/overrides`, label: "Translation rules" },
+    { href: `${baseHref}/admin`, label: "Settings" },
   ];
 
   return (
