+++ app/dashboard/sites/page.tsx
@@ -1,113 +1,5 @@
-import Link from "next/link";
-import { headers } from "next/headers";
-
-import { Button } from "@/components/ui/button";
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { ErrorStateCard } from "@/components/dashboard/error-state-card";
-import { listSitesCached } from "@internal/dashboard/data";
-import { resolveDashboardErrorView } from "@internal/dashboard/error-state";
-import { type SiteSummary } from "@internal/dashboard/webhooks";
-import { requireDashboardAuth } from "@internal/dashboard/auth";
-import { resolvePreferredLocale } from "@internal/i18n";
-import { SitesList } from "../_components/sites-list";
+import { redirect } from "next/navigation";
 
 export default async function SitesPage() {
-  let sites: SiteSummary[] = [];
-  let error: unknown = null;
-  let canCreateSite = false;
-  let billingBlocked = false;
-  let atSiteLimit = false;
-  const locale = resolvePreferredLocale((await headers()).get("accept-language"));
-  const pricingPath = `/${locale}/pricing`;
-
-  try {
-    const auth = await requireDashboardAuth();
-    sites = await listSitesCached(auth.webhooksAuth!);
-    billingBlocked = !auth.mutationsAllowed;
-    const maxSites = auth.account?.featureFlags.maxSites ?? null;
-    const activeSites = sites.filter((site) => site.status === "active").length;
-    const hasAvailableSlot = maxSites === null || activeSites < maxSites;
-    atSiteLimit = maxSites !== null && activeSites >= maxSites;
-    canCreateSite = auth.has({ feature: "site_create" }) && !billingBlocked && hasAvailableSlot;
-  } catch (err) {
-    error = err;
-  }
-
-  const errorView = resolveDashboardErrorView(error, {
-    title: "Could not load sites",
-    description:
-      "We could not complete your request. You can retry or return to the dashboard home.",
-    message: "Unable to load sites.",
-  });
-
-  return (
-    <div className="space-y-6">
-      <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-        <div className="space-y-1">
-          <h2 className="text-2xl font-semibold">Sites</h2>
-          <p className="text-sm text-muted-foreground">
-            Review onboarding status, domains, and languages for every property.
-          </p>
-        </div>
-        {error ? null : canCreateSite ? (
-          <Button asChild>
-            <Link href="/dashboard/sites/new" title="Add a site">
-              Add a site
-            </Link>
-          </Button>
-        ) : billingBlocked ? (
-          <Button asChild variant="secondary">
-            <Link href={pricingPath} title="Update billing">
-              Update billing
-            </Link>
-          </Button>
-        ) : atSiteLimit ? (
-          <div className="flex flex-wrap items-center gap-2">
-            <Button disabled title="Site limit reached">
-              Add a site
-            </Button>
-            <Button asChild variant="secondary">
-              <Link href={pricingPath} title="Upgrade for more sites">
-                Upgrade for more sites
-              </Link>
-            </Button>
-          </div>
-        ) : (
-          <Button asChild variant="secondary">
-            <Link href={pricingPath} title="Upgrade to add a site">
-              Upgrade to add a site
-            </Link>
-          </Button>
-        )}
-      </div>
-
-      {error ? (
-        <ErrorStateCard
-          title={errorView.title}
-          description={errorView.description}
-          message={errorView.message}
-        />
-      ) : sites.length === 0 ? (
-        <Card>
-          <CardHeader>
-            <CardTitle>No sites yet</CardTitle>
-            <CardDescription>
-              Start with the onboarding wizard to set up your first site.
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            {canCreateSite ? (
-              <Button asChild>
-                <Link href="/dashboard/sites/new" title="Start onboarding">
-                  Start onboarding
-                </Link>
-              </Button>
-            ) : null}
-          </CardContent>
-        </Card>
-      ) : (
-        <SitesList sites={sites} />
-      )}
-    </div>
-  );
+  redirect("/dashboard");
 }
